requirements.txt
fastapi
uvicorn
requests
networkx
pydantic
pytest
httpx
clerk-backend-api  
python-jose         